+++ cmd/template/cluster/provider/aks.go
@@ -1,0 +1,158 @@
+package provider
+
+import (
+	"context"
+	"io"
+	"text/template"
+
+	"github.com/giantswarm/k8sclient/v8/pkg/k8sclient"
+	"github.com/giantswarm/k8smetadata/pkg/label"
+	"github.com/giantswarm/microerror"
+	"sigs.k8s.io/yaml"
+
+	"github.com/giantswarm/kubectl-gs/v6/cmd/template/cluster/common"
+	"github.com/giantswarm/kubectl-gs/v6/cmd/template/cluster/provider/templates/aks"
+	"github.com/giantswarm/kubectl-gs/v6/internal/key"
+	templateapp "github.com/giantswarm/kubectl-gs/v6/pkg/template/app"
+)
+
+const (
+	ClusterAKSRepoName = "cluster-aks"
+	ReleaseAKSRepoName = "release-aks"
+)
+
+func WriteAKSTemplate(ctx context.Context, client k8sclient.Interface, output io.Writer, config common.ClusterConfig) error {
+	err := templateClusterAKS(ctx, client, output, config)
+	if err != nil {
+		return microerror.Mask(err)
+	}
+
+	return nil
+}
+
+func templateClusterAKS(ctx context.Context, k8sClient k8sclient.Interface, output io.Writer, config common.ClusterConfig) error {
+	appName := config.Name
+	configMapName := common.UserConfigMapName(appName)
+
+	var configMapYAML []byte
+	{
+		flagValues := BuildAKSClusterConfig(config)
+
+		// For release versions, the release version is baked into the chart,
+		// so we don't need to include it in the user config.
+		if common.IsReleaseVersion(config.ReleaseVersion) {
+			flagValues.Global.Release = nil
+		}
+
+		configData, err := aks.GenerateClusterValues(flagValues)
+		if err != nil {
+			return microerror.Mask(err)
+		}
+
+		userConfigMap, err := templateapp.NewConfigMap(templateapp.UserConfig{
+			Name:      configMapName,
+			Namespace: common.OrganizationNamespace(config.Organization),
+			Data:      configData,
+		})
+		if err != nil {
+			return microerror.Mask(err)
+		}
+
+		userConfigMap.Labels = map[string]string{}
+		userConfigMap.Labels[label.Cluster] = config.Name
+		for k, v := range config.Labels {
+			userConfigMap.Labels[k] = v
+		}
+		if config.PreventDeletion {
+			userConfigMap.Labels[label.PreventDeletion] = "true" //nolint:goconst
+		}
+
+		configMapYAML, err = yaml.Marshal(userConfigMap)
+		if err != nil {
+			return microerror.Mask(err)
+		}
+	}
+
+	var appYAML []byte
+	{
+		// Use release-<provider> chart name for release versions (>= 35.0.0).
+		// These charts have the release version baked into values.yaml.
+		// For older chart versions, use cluster-<provider> and let the webhook handle version.
+		chartName := ClusterAKSRepoName
+		if common.IsReleaseVersion(config.ReleaseVersion) {
+			chartName = ReleaseAKSRepoName
+		}
+
+		clusterAppConfig := templateapp.Config{
+			AppName:                 config.Name,
+			Catalog:                 config.App.ClusterCatalog,
+			InCluster:               true,
+			Name:                    chartName,
+			Namespace:               common.OrganizationNamespace(config.Organization),
+			UserConfigConfigMapName: configMapName,
+			ExtraLabels:             map[string]string{},
+		}
+		// Set version for release charts where the chart version equals the release version.
+		// For cluster-<provider> charts, the webhook handles version mutation.
+		if common.IsReleaseVersion(config.ReleaseVersion) {
+			clusterAppConfig.Version = config.ReleaseVersion
+		}
+		for k, v := range config.Labels {
+			clusterAppConfig.ExtraLabels[k] = v
+		}
+		if config.PreventDeletion {
+			clusterAppConfig.ExtraLabels[label.PreventDeletion] = "true"
+		}
+
+		var err error
+		appYAML, err = templateapp.NewAppCR(clusterAppConfig)
+		if err != nil {
+			return microerror.Mask(err)
+		}
+	}
+
+	t := template.Must(template.New("appCR").Parse(key.AppCRTemplate))
+
+	err := t.Execute(output, templateapp.AppCROutput{
+		AppCR:               string(appYAML),
+		UserConfigConfigMap: string(configMapYAML),
+	})
+	return microerror.Mask(err)
+}
+
+func BuildAKSClusterConfig(config common.ClusterConfig) aks.ClusterConfig {
+	providerSpecific := &aks.ProviderSpecific{
+		Location:       config.Region,
+		SubscriptionID: config.Azure.SubscriptionID,
+	}
+	// Only emit the asoAuthentication block when the user supplied workload-identity
+	// values, otherwise rely on the chart's default credential resolution.
+	if config.AKS.TenantID != "" || config.AKS.ClientID != "" {
+		providerSpecific.ASOAuthentication = &aks.ASOAuthentication{
+			SubscriptionID: config.Azure.SubscriptionID,
+			TenantID:       config.AKS.TenantID,
+			ClientID:       config.AKS.ClientID,
+		}
+	}
+
+	return aks.ClusterConfig{
+		Global: &aks.Global{
+			Metadata: &aks.Metadata{
+				Name:            config.Name,
+				Description:     config.Description,
+				Labels:          config.Labels,
+				Organization:    config.Organization,
+				PreventDeletion: config.PreventDeletion,
+			},
+			ControlPlane: &aks.ControlPlane{
+				SKU: &aks.SKU{
+					Tier: "Standard",
+				},
+			},
+			ProviderSpecific: providerSpecific,
+			Release: &aks.Release{
+				Version: config.ReleaseVersion,
+			},
+		},
+	}
+}